requirements.txt
absl-py==0.13.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
distlib==0.3.2
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
imageio==2.9.0
ipython @ file:///C:/ci/ipython_1617121002983/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.3.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-docs===0.0.047a807024a49d648ff635434d49e3f03a8cfc57a-
tensorflow-estimator==2.5.0
termcolor==1.1.0
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.7.4.3
urllib3==1.26.6
virtualenv==20.4.7
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
